refactor(mentor): 폴더명과 컴포넌트명 일치화 (Prodcut)
